build_bsp: Fix u-boot signing issue
TI_SECURE_DEV_PKG has to be exported for uboot and k3ig makefiles to
be able to use it. Apparently, setting the path in env is not sufficient
and "export" is needed. One other way to provide this is add
TI_SECURE_DEV_PKG=${TI_SECURE_DEV_PKG} as a prepend to make commands.

Signed-off-by: Sai Sree Kartheek Adivi <sskartheek.adivi@outlook.com>