Verify the validity of the build configuration before building
Presently, the script does not check whether the build exists before
calling bdebstrap. This leads to it failing messily later in the
process.

To fix this, first verify that the build itself exists in builds.toml
file. Then verify that the machine, the bsp_version and the
distro_variant also exist, in configs/machines.toml,
configs/bsp_sources.toml and configs/bdebstrap/ respectively.

Signed-off-by: Suhaas Joshi <joshiesuhaas0@gmail.com>