bdebstrap-configs: Add new configs for building in chroot env
Out of tree kernel modules use DKMS tool in the deb package.
DKMS builds the kernel modules for the kernel that's currently installed.

bdebstrap/mmdebstrap uses chroot env to build the rootfs. In a chroot
environment, `uname -r` gives host machines kernel version but not the
one installed in the chroot environment since it's not actually booted.

So DKMS tries to find headers for wrong version of kernel. So create a
bdebstrap config 'bookworm-default-no-km.yaml' which will not install
any out of tree kernel modules. These has to be installed once after the
image is booted.

The alternative is to build on the target itself which has proper kernel
version already installed using bookworm-default.yaml.

---

Also update the builds.toml to use bookworm-default-no-km instead of
bookworm-default because github workflows run on hosts provided by
Github and they might not be of arm64 architecture and will not have
expected kernel variant/version.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>