create-wic: Add a script to create SD Card image
The script that's available now (create-sdcard.sh) can be used to format
and flash the build to a physical SD card.

But we need an Image generation script so that a build can be shared and
flashed on SD Cards with more standard tools like balena-etcher. This
new script provides that functionality.

Signed-off-by: Sai Sree Kartheek Adivi <sskartheek.adivi@outlook.com>