Fix LPM: Return and use array instead of just 1st string
A config in toml can have multiple values, separated by space. The
`read_config` function reads all of these space-separated values as
individual strings in a `value` array. But then it returns only the
first string by doing `echo ${value}`. Thus even if a config had
multiple values, only the 1st gets returned.

This is usually not a problem since almost all of the configs so far
have single-values. But in case of optee's make arguments, which are
multiple, it returns only the 1st, and not the 2nd. Due to this, LPM
doesn't work on am62pxx-evm, since the 2nd make argument is also
necessary for making it work.

So return the entire array from `read_config`. And pass the entire array
to optee's `make` command. Also fix the value of the `platform` config
for am62pxx-evm.

This commit fixes LPM.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>