Change network interface names to endX from ethX
Systemd uses "Predictable Network Interface Names", which assigns names
such as:
    * ethX -> legacy naming (shouldn't be used for new systems, the
      Debian docs tell)
    * ensX -> PCI devices
    * endX -> for ARM-based embedded devices

Earlier, we were able to use ethX, but with 6.12 kernel and systemd
changes, I guess this is no longer straightforward. So change these
names to endX (that is, eth0 -> end0 and eth1 -> end1).

Tested wifi with this commit too.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>